refactor(postDiaryEntryTest): refactor diary entry test folders/files to make it more modular
